left align rename of controller to associate with activities
