+++ backend/index.js
@@ -18,7 +18,7 @@ const authentication = require('./controllers/authentication');
 const userCaseLoads = require('./controllers/usercaseloads');
 const setActiveCaseLoad = require('./controllers/setactivecaseload');
 const userLocations = require('./controllers/userLocations');
-const locations = require('./controllers/locations');
+const locations = require('./controllers/activities');
 const userMe = require('./controllers/userMe');
 const getConfig = require('./controllers/getConfig');
 const houseblockLocations = require('./controllers/houseblockLocations');
