+++ src/Dropdown/theme.scss
@@ -5,7 +5,7 @@
   background: $govuk-blue;
   display: flex;
   height: 100%;
-  min-width: 269px;
+  min-width: 320px;
   position: relative;
 }
 
@@ -50,7 +50,7 @@
   color: white;
   font-size: 19px;
   align-items: center;
-  text-align: center;
+  text-align: left;
   border-top: 1px solid lightblue;
   height: 60px;
   text-decoration: none;
@@ -69,7 +69,7 @@
   color: white;
   font-size: 19px;
   align-items: center;
-  text-align: center;
+  text-align: left;
   border-top: 1px solid lightblue;
   height: 60px;
   text-decoration: none;
